dbricks_helpers/demos/azure/databricks_apps/data-contract-authoring-app/requirements.txt
psycopg2-binary>=2.9.9
databricks-sdk>=0.60.0
pyyaml>=6.0.1
tabulate>=0.9.0
streamlit>=1.35.0
pyyaml
pyspark
streamlit_ace